+++ rwastateview.cpp
@@ -454,17 +454,20 @@ void RwaStateView::addAssetItem(const QString &path, qint32 type)
         TagLib::FileRef f(path.toStdString().c_str());
         int channels = 0;
         int length = 0;
+        int sampleRate = 0;
 
         if(!f.isNull() && f.audioProperties())
         {
             channels = f.audioProperties()->channels();
             length = f.audioProperties()->lengthInMilliseconds();
+            sampleRate = f.audioProperties()->sampleRate();
         }
 
         string uid = std::string(QUuid::createUuid().toString().toLatin1());
         RwaAsset1 *newItem = new RwaAsset1(path.toStdString(), currentState->getCoordinates(), type, uid);
         newItem->setDuration(length);
         newItem->setNumberOfChannels(channels);
+        newItem->setOriginalSampleRate(sampleRate);
         currentState->addAsset(newItem);
         currentState->setLastTouchedAsset(newItem);
         emit sendCurrentState(currentState);